http://jira.aplana.com/browse/SBRFACCTAX-12116 0.6.1 Неверный расчет по "длинным цифрам" найден в декларации НДС. Просмотреть скрипты всех НФ и деклараций для поиска такой же проблемы.

+++ src/main/resources/form_template/transport/declaration/v2012/script.groovy
@@ -277,10 +277,10 @@ def bildXml(def departmentParamTransport, def formDataCollection, def department
                             СумПУ(
                                     ОКТМО: getOkato(okato),
                                     НалИсчисл: row.taxSumToPay,
-                                    АвПУКв1: row.amountOfTheAdvancePayment1.setScale(0, BigDecimal.ROUND_HALF_UP).intValue(),
-                                    АвПУКв2: row.amountOfTheAdvancePayment2.setScale(0, BigDecimal.ROUND_HALF_UP).intValue(),
-                                    АвПУКв3: row.amountOfTheAdvancePayment3.setScale(0, BigDecimal.ROUND_HALF_UP).intValue(),
-                                    НалПУ: row.amountOfTaxPayable.setScale(0, BigDecimal.ROUND_HALF_UP).intValue(),
+                                    АвПУКв1: row.amountOfTheAdvancePayment1.setScale(0, BigDecimal.ROUND_HALF_UP).longValue(),
+                                    АвПУКв2: row.amountOfTheAdvancePayment2.setScale(0, BigDecimal.ROUND_HALF_UP).longValue(),
+                                    АвПУКв3: row.amountOfTheAdvancePayment3.setScale(0, BigDecimal.ROUND_HALF_UP).longValue(),
+                                    НалПУ: row.amountOfTaxPayable.setScale(0, BigDecimal.ROUND_HALF_UP).longValue(),
                             ) {
                                 row.rowData.each { tRow ->
                                     def taxBenefitCode = tRow.taxBenefitCode ? getRefBookValue(6, tRow.taxBenefitCode)?.CODE?.stringValue : null
